refactor: Merge types/docs from testing/auth/base.pyi
- Merged type hints and docstrings from `crudclient/testing/auth/base.pyi` into `crudclient/testing/auth/base.py`.

- Deleted `crudclient/testing/auth/base.pyi`.
